Song 2015-10-23 1
GeneralPreference 클래스 추가 및 세션 관련 기능 추가

페이스북과 앱 연동하여 세션유지하기 위해
