+++ app/src/main/java/Items/StaticVariable.java
@@ -7,12 +7,12 @@ public class StaticVariable {
 
 
     private static boolean userLogined = false;
-    private static String user_id = "123";
-    private static String user_name = "temp12312";
+    private static String user_id = "NULL";
+    private static String user_name = "NULL";
 
     private static String connectUrl = "https://220.72.126.199:15443/";
 
-    public static boolean isUserLogined() {
+    public static boolean getUserLogined() {
         return userLogined;
     }
