+++ app/src/main/res/layout/fragment_navigation_drawer.xml
@@ -1,6 +1,6 @@
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="Items.NavigationDrawerFragment"
+    tools:context="com.example.dong_gyo.project.NavigationDrawerFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical">
@@ -15,7 +15,7 @@
         <TextView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:id="@+id/userInfoTitle"
+            android:id="@+id/show_user_name"
             android:textSize="20sp"
             android:text="내정보"
             />
@@ -25,6 +25,12 @@
             android:layout_height="wrap_content"
             android:text="로그인"
             android:id="@+id/mv_loginbutton"/>
+
+        <Button
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="내정보"
+            android:id="@+id/mv_myinfobutton"/>
 
     </LinearLayout>
