+++ local.properties
@@ -7,5 +7,5 @@
 # Location of the SDK. This is only used by Gradle.
 # For customization when using a Version Control System, please read the
 # header note.
-#Wed Oct 14 22:00:09 KST 2015
+#Fri Oct 23 13:56:39 KST 2015
 sdk.dir=D\:\\Song\\AppData\\Local\\Android\\sdk
